+++ backend/server.py
@@ -44,6 +44,14 @@ token_manager = TokenManager()
 app = Flask(__name__)
 CORS(app) # allow requests from desktop device browsers
 
+@app.after_request # runs after every request
+# This method adds CORS headers even when responses are unsuccessful to prevent CORS errors when trying to debug
+def add_cors_headers(response):
+    response.headers['Access-Control-Allow-Origin'] = 'https://nicholasdakis.github.io'
+    response.headers['Access-Control-Allow-Headers'] = 'Content-Type,Authorization'
+    response.headers['Access-Control-Allow-Methods'] = 'GET,POST,OPTIONS'
+    return response
+
 # Initialize Firestore client
 # Load credentials JSON string from env var
 cred_json = os.environ.get("GOOGLE_APPLICATION_CREDENTIALS_JSON")